move overlap logic to hook
