+++ app/Schengen/Schengen.tsx
@@ -5,13 +5,7 @@ import { useQueryDates } from "app/DaysTracker/useQueryDates";
 import { DayActions } from "app/DaysTracker/Day/DayActions";
 import { Menu } from "app/DaysTracker/Menu";
 import { SchengenTrip } from "app/Schengen/SchengenTrip";
-import {
-  areIntervalsOverlapping,
-  isAfter,
-  isPast,
-  isValid,
-  isWithinInterval,
-} from "date-fns";
+import { isAfter, isPast } from "date-fns";
 import { Summary } from "app/Schengen/Summary";
 import { SchengenDate } from "app/Schengen/types";
 import { AddSchengenDate } from "app/Schengen/AddSchengenDate";
@@ -54,42 +48,6 @@ export const Schengen = () => {
   };
   const resetDates = () => setDates([]);
 
-  const overlappingTrips = (date: Date | string, endDate?: Date | string) => {
-    if (!date && !endDate) return [];
-
-    if (endDate) {
-      if (!isValid(new Date(endDate)) || !isValid(new Date(date))) {
-        console.error(
-          `received invalid dates ("${date}", "${endDate}") when trying to find overlapping trips.`,
-        );
-        return [];
-      }
-      return dates().filter((trip) =>
-        areIntervalsOverlapping(
-          { start: date, end: endDate },
-          {
-            start: trip.date,
-            end: trip.endDate,
-          },
-        ),
-      );
-    }
-
-    if (!isValid(new Date(date))) {
-      console.error(
-        `received invalid date "${date}" when trying to find overlapping trips.`,
-      );
-      return [];
-    }
-
-    return dates().filter((trip) =>
-      isWithinInterval(date, {
-        start: trip.date,
-        end: trip.endDate,
-      }),
-    );
-  };
-
   const fallback = <p class="text-gray-500">Add a trip below</p>;
 
   return (
@@ -120,12 +78,8 @@ export const Schengen = () => {
           )}
         </For>
       </ul>
-      <AddSchengenDate
-        dates={dates}
-        overlappingTrips={overlappingTrips}
-        setDates={setDates}
-      />
-      <Summary dates={dates} overlappingTrips={overlappingTrips} />
+      <AddSchengenDate dates={dates} setDates={setDates} />
+      <Summary dates={dates} />
       <Menu dates={dates} resetDates={resetDates} />
     </main>
   );
